requirements.txt
phidata 
python-dotenv
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
openbb
polars
pyarrow
python-docx